requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
google-generativeai
pytest
httpx
pytest-asyncio
black
faiss-cpu
sentence-transformers
PyPDF2
pdfplumber
pandas
hf_xet